fix(review): P0/P2 — stats SSOT, lorem, testimonial, bio, atproto, cadence
- single source of truth for headline numbers (src/data/siteStats.ts:
  yearsBuilding, countryCount, talksCount); homeContent re-exports it
- remove the dead lorem-ipsum HTML comment in the gallery + restyle its header
  to the redesign system (drop old Badge/.h2)
- testimonials: Chris Rauch is CCO at Supermetrics (was Spryker)
- author bio: current identity, not "Evangelist at Spryker"
- newsletter cadence: "a few times a year" (was "monthly highlight reel")
- atproto: fix ATPROTO/ATProtocol typos, add reciprocal rel="me" to Bluesky
