+++ src/content/authors/gxjansen/index.mdx
@@ -5,9 +5,9 @@ email: x@gui.do
 authorLink: https://gui.do
 
 # Professional Identity
-jobTitle: Global Business & Technology Evangelist
-organization: Spryker
-organizationUrl: https://spryker.com/
+jobTitle: Community builder & cognitive psychologist
+organization: Singi Labs
+organizationUrl: https://gui.do
 
 # Biography
 bioShort: Community strategist with 20+ years building technical ecosystems around open-source platforms.
