+++ src/content/post/introducing-barazo-community-forums/index.mdx
@@ -32,7 +32,7 @@ These patterns bugged me enough to try building something different. (Apparently
 <a href="https://barazo.forum">
   <Image
     src={barazoBanner}
-    alt="Barazo banner: Your identity. Your communities. Everywhere. Built on ATProtocol."
+    alt="Barazo banner: Your identity. Your communities. Everywhere. Built on the AT Protocol."
   />
 </a>
