+++ src/layouts/BaseHead.astro
@@ -135,8 +135,9 @@ const noindex = !CONFIG.ALLOW_INDEXING;
 <meta name="generator" content={Astro.generator} />
 <link rel="sitemap" href="/sitemap-index.xml" />
 <link rel="me" href="https://sifa.id/p/gui.do" />
+<link rel="me" href="https://bsky.app/profile/gui.do" />
 
-{/* Standard.site document verification (ATProto long-form publishing) */}
+{/* Standard.site document verification (AT Protocol long-form publishing) */}
 {
   standardSiteConfig.enabled &&
     type === "blog" &&
