+++ src/pages/newsletter.astro
@@ -44,7 +44,7 @@ import NewsletterForm from "@components/Newsletter/NewsletterForm.astro";
 
         <div class="prose dark:prose-invert mx-auto mt-10 max-w-none">
           <h2>What you get</h2>
-          <p>A monthly highlight reel, scannable and focused:</p>
+          <p>A few times a year, scannable and focused:</p>
           <ul>
             <li>
               <strong>Big Tech alternatives:</strong> privacy-respecting tools for
